Alterado o tipo do campo point

+++ ZSSN/src/main/java/br/com/zssn/entity/Resource.java
@@ -34,5 +34,5 @@ public class Resource {
 	private String name;
 
 	@Column(nullable = false)
-	private Double point;
+	private Integer point;
 }
